refactor: common.domain 패키지의 Name, Price, Quantity를 메뉴 패키지로 이동
