+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -11,8 +11,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.common.domain.Name;
-import kitchenpos.common.domain.Price;
 
 @Entity
 public class Menu {
