+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -5,7 +5,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.common.domain.Name;
 
 @Entity
 public class MenuGroup {
