+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -10,7 +10,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.common.domain.Quantity;
 
 @Entity
 public class MenuProduct {
