+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -6,7 +6,6 @@ import java.util.Collections;
 import java.util.List;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import kitchenpos.common.domain.Price;
 
 @Embeddable
 public class MenuProducts {
