+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -6,8 +6,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.common.domain.Name;
-import kitchenpos.common.domain.Price;
 
 @Entity
 public class Product {
