+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -10,7 +10,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.common.domain.Quantity;
+import kitchenpos.menu.domain.Quantity;
 
 @Entity
 public class OrderLineItem {
