+++ src/test/java/kitchenpos/menu/domain/PriceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.common.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import kitchenpos.menu.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
