+++ src/test/java/kitchenpos/menu/domain/QuantityTest.java
@@ -1,8 +1,8 @@
-package kitchenpos.common.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import kitchenpos.common.domain.Quantity;
+import kitchenpos.menu.domain.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
